[scroll-animations] Recalculate siblings if needed
Due to how the named timeline lookup works, it is now possible for
the computed style of an element to depend on the style of a previous
sibling. This requires a new kind of StyleRecalcChange::Propagation
which can "upgrade" the style recalc of an element to also include
the subsequent siblings (and their descendants).

In order to do this, Element::RecalcStyle now returns a
StyleRecalcChange which is intended to be combined with an outer
StyleRecalcChange to "upgrade" the recalc severity at the call-site.

 - If we are already in ContainerNode::RecalcDescendantStyles, then
   that's effectively a sibling traversal from the perspective of a
   given child. So we can piggy-back on that traversal in this case.
 - Otherwise, if the sibling-recalc-triggering element is the recalc
   root itself, the new ContainerNode::RecalcSubsequentSiblingStyles
   function will handle it.

Note that the following things are deliberately ignored in this CL:

 - ShadowDOM
 - Forced style updates (NeedsLayoutTreeUpdateForNode)

Bug: 1317765
Change-Id: Ib78d8815449318420d6aab09dc671a2b5f0a168f